* move all gradient methods to extension * add underlay to single color elements to compensate density
